moved SnapshotRepository to core, added addResourceForEntity with a SnapShotRepository as last parameter
