+++ crabzilla-core/src/main/java/io/github/crabzilla/core/concrete.kt
@@ -1,7 +1,14 @@
 package io.github.crabzilla.core
 
-import kotlinx.serialization.PolymorphicSerializer
+import io.github.crabzilla.internal.CommandController
+import io.vertx.core.Future
+import io.vertx.core.Promise
+import io.vertx.core.json.JsonObject
+import io.vertx.core.shareddata.SharedData
 import java.util.UUID
+import kotlinx.serialization.PolymorphicSerializer
+import kotlinx.serialization.json.Json
+import org.slf4j.LoggerFactory
 
 val ENTITY_SERIALIZER = PolymorphicSerializer(Entity::class)
 val COMMAND_SERIALIZER = PolymorphicSerializer(Command::class)
@@ -60,5 +67,73 @@ data class UnitOfWork(
     const val COMMAND = "command"
     const val VERSION = "version"
     const val EVENTS = "events"
+  }
+}
+
+/**
+ * A light implementation
+ */
+class InMemorySnapshotRepository<E : Entity>(
+  private val sharedData: SharedData,
+  private val json: Json,
+  private val entityName: String,
+  private val initialState: E
+) : SnapshotRepository<E> {
+  companion object {
+    internal val log = LoggerFactory.getLogger(CommandController::class.java)
+  }
+  override fun upsert(entityId: Int, snapshot: Snapshot<E>): Future<Void> {
+    val promise = Promise.promise<Void>()
+    sharedData.getAsyncMap<Int, String>(entityName) { event1 ->
+      if (event1.failed()) {
+        log.error("Failed to get map $entityName")
+        promise.fail(event1.cause())
+        return@getAsyncMap
+      }
+      val stateAsJson = JsonObject(json.stringify(ENTITY_SERIALIZER, snapshot.state))
+      val mapEntryAsJson = JsonObject().put("version", snapshot.version).put("state", stateAsJson)
+      event1.result().put(entityId, mapEntryAsJson.encode()) { event2 ->
+        if (event2.failed()) {
+          log.error("Failed to put $entityId on map $entityName")
+          promise.fail(event2.cause())
+          return@put
+        }
+        promise.complete()
+      }
+    }
+    return promise.future()
+  }
+
+  override fun retrieve(entityId: Int): Future<Snapshot<E>> {
+    val promise = Promise.promise<Snapshot<E>>()
+    val defaultSnapshot = Snapshot(initialState, 0)
+    sharedData.getAsyncMap<Int, String>(entityName) { event1 ->
+      if (event1.failed()) {
+        log.error("Failed get map $entityName")
+        promise.fail(event1.cause())
+        return@getAsyncMap
+      }
+      event1.result().get(entityId) { event2 ->
+        if (event2.failed()) {
+          log.error("Failed to get $entityId on map $entityName")
+          promise.fail(event2.cause())
+          return@get
+        }
+        val result = event2.result()
+        if (result == null) {
+          if (log.isDebugEnabled) {
+            log.debug("Returning default snapshot for $entityId on map $entityName")
+          }
+          promise.complete(defaultSnapshot)
+          return@get
+        }
+        val mapEntryAsJson = JsonObject(event2.result())
+        val version = mapEntryAsJson.getInteger("version")
+        val stateAsJson = mapEntryAsJson.getJsonObject("state")
+        val state = json.parse(ENTITY_SERIALIZER, stateAsJson.encode()) as E
+        promise.complete(Snapshot(state, version))
+      }
+    }
+    return promise.future()
   }
 }
