+++ crabzilla-core/src/test/java/io/github/crabzilla/UnitOfWorkTest.kt
@@ -4,10 +4,10 @@ import io.github.crabzilla.core.DomainEvent
 import io.github.crabzilla.core.UnitOfWork
 import io.github.crabzilla.example1.CreateCustomer
 import io.github.crabzilla.example1.CustomerCreated
+import java.util.UUID
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Assertions.assertThrows
 import org.junit.jupiter.api.Test
-import java.util.UUID
 
 class UnitOfWorkTest {
