+++ crabzilla-pg-client-web/src/test/java/io/github/crabzilla/pgc/web/example1/CustomerVerticle.kt
@@ -1,8 +1,9 @@
 package io.github.crabzilla.pgc.web.example1
 
-import io.github.crabzilla.pgc.web.WebResourceContext
+import io.github.crabzilla.core.InMemorySnapshotRepository
 import io.github.crabzilla.pgc.web.PgcReadContext
 import io.github.crabzilla.pgc.web.PgcWriteContext
+import io.github.crabzilla.pgc.web.WebResourceContext
 import io.github.crabzilla.pgc.web.addProjector
 import io.github.crabzilla.pgc.web.addResourceForEntity
 import io.github.crabzilla.pgc.web.listenHandler
@@ -47,7 +48,11 @@ class CustomerVerticle : AbstractVerticle() {
 
     val writeContext = PgcWriteContext(vertx, example1Json, writeDb)
     val resourceContext = WebResourceContext("customers", "customer", CustomerCommandAware(), cmdTypeMap)
-    addResourceForEntity(writeContext, resourceContext, router)
+    // addResourceForEntity(writeContext, resourceContext, router)
+
+    // or
+    val inMemoryRepo = InMemorySnapshotRepository(vertx.sharedData(), example1Json, "customer", Customer())
+    addResourceForEntity(writeContext, resourceContext, router, inMemoryRepo)
 
     // projection consumers
     val readContext = PgcReadContext(vertx, example1Json, readDb)
