+++ crabzilla-pg-client/src/test/java/io/github/crabzilla/pgc/PgcSnapshotRepoIT.kt
@@ -3,7 +3,7 @@ package io.github.crabzilla.pgc
 import io.github.crabzilla.core.COMMAND_SERIALIZER
 import io.github.crabzilla.core.EVENT_SERIALIZER
 import io.github.crabzilla.core.Snapshot
-import io.github.crabzilla.internal.SnapshotRepository
+import io.github.crabzilla.core.SnapshotRepository
 import io.github.crabzilla.pgc.PgcUowJournal.Companion.SQL_APPEND_UOW
 import io.github.crabzilla.pgc.example1.Customer
 import io.github.crabzilla.pgc.example1.CustomerCommandAware
@@ -25,6 +25,7 @@ import io.vertx.junit5.VertxExtension
 import io.vertx.junit5.VertxTestContext
 import io.vertx.pgclient.PgPool
 import io.vertx.sqlclient.Tuple
+import java.util.UUID
 import kotlinx.serialization.builtins.list
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
@@ -32,7 +33,6 @@ import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.TestInstance
 import org.junit.jupiter.api.extension.ExtendWith
-import java.util.UUID
 
 @ExtendWith(VertxExtension::class)
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
